refactor `Header` component
