--- src/components/Edit/Header.jsx
@@ -1,40 +1,0 @@
-import { useState, useEffect } from "react";
-
-export default function Header() {
-	return (
-		<>
-			<header className="flex flex-col items-center gap-1 w-full">
-				<h1 className="patrick-hand-sc-regular text-6xl text-yellow-700">
-					Edit Price
-				</h1>
-				<h2 className="patrick-hand-regular text-lg text-yellow-700 tracking-wider">
-					<SubHeading />
-				</h2>
-			</header>
-		</>
-	);
-}
-
-function SubHeading() {
-	const subHeadings = ["Enter custom price", "Prices update instantly"];
-
-	const [currentSubheadingIndex, setCurrentSubheadingIndex] = useState(0);
-
-	// Function to cycle through subheadings
-	const changeSubheading = () => {
-		setCurrentSubheadingIndex(
-			(prevIndex) => (prevIndex + 1) % subHeadings.length
-		);
-	};
-
-	// Change subheading every 5 seconds
-	useEffect(() => {
-		const intervalId = setInterval(() => {
-			changeSubheading();
-		}, 2500); // Change subheading every 2.5 seconds
-
-		return () => clearInterval(intervalId);
-	});
-
-	return subHeadings[currentSubheadingIndex];
-}
