+++ src/pages/Edit.jsx
@@ -1,5 +1,5 @@
 import { ValidateItemsStateProps } from "./../propValidations";
-import Header from "./../components/Edit/Header";
+import Header from "../components/Header";
 import ScrapItems from "../components/Edit/ScrapItems";
 import Footer from "./../components/Footer";
 import Aside from "./../components/Aside";
@@ -8,9 +8,11 @@ import homeIcon from "./../assets/home.svg";
 Edit.propTypes = ValidateItemsStateProps;
 
 function Edit({ items, setItems }) {
+	const subHeadings = ["Enter custom price", "Prices update instantly"];
+
 	return (
 		<>
-			<Header />
+			<Header title="Edit Price" subHeadings={subHeadings} />
 			<Aside icon={homeIcon} alt="home" linkTo="/" />
 			<ScrapItems items={items} setItems={setItems} />
 			<Footer />
